Don't force importing modules only used when running as a script
The sys and argparse module are never used in any of the useful
functions of fedora_active_user. They are only used if running it as a
script.

But there's a bunch of interesting stuff which could make
fedora_active_user attractive as a module to import.

Only importing sys and argparse where they are actually used avoids
forcing them on people using fedora_active_user as a module.